Removed dependencies and moved code as example.
