+++ build.gradle
@@ -49,7 +49,6 @@ dependencies {
   compile 'com.fasterxml.jackson.core:jackson-core:2.5.3'
   compile 'com.fasterxml.jackson.core:jackson-annotations:2.5.3'
   compile 'com.fasterxml.jackson.core:jackson-databind:2.5.3'
-  compile 'com.dropbox.core:dropbox-core-sdk:3.0.5'
   testCompile group: 'junit', name: 'junit', version: '4.12'
 }
