+++ src/main/java/com/sendgrid/helpers/mail/objects/Attachments.java
@@ -242,23 +242,4 @@ public class Attachments {
       return attachments;
     }
   }
-
-  /**
-   * Uploads attachment to Dropbox
-   *
-   * @param accessToken user's Dropbox access token
-   * @param path path of the folder in which attachment needs to be added. Should not end with /
-     * @return the full path to the uploaded file
-     */
-  public String uploadToDropbox(String accessToken, String path){
-    try {
-      DbxRequestConfig config = DbxRequestConfig.newBuilder("sendgrid/0.1").build();
-      DbxClientV2 client = new DbxClientV2(config, accessToken);
-      FileMetadata uploadedFile = client.files().upload(String.format("%s/%s", path, filename))
-              .uploadAndFinish(new ByteArrayInputStream((Base64.decodeBase64(content))));
-      return uploadedFile.getPathDisplay();
-    }catch(Exception ex) {
-      throw new RuntimeException("Error while uploading to Dropbox", ex);
-    }
-  }
 }
